backend/requirements.txt
fastapi
uvicorn
python-dotenv
deepgram-sdk
groq
pydantic
python-multipart==0.0.9
requests
